LayoutTests/platform/wpe/fast/canvas/webgl/vertexAttribPointer-with-bad-offset-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_VALUE: vertexAttribPointer: bad offset
CONSOLE MESSAGE: WebGL: INVALID_OPERATION: drawArrays: no buffer is bound to enabled attribute
PASS: vertexAttribPointer should have an error.
